# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Swagger Petstore",
        "description": "This is a sample server Petstore server.  You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).  For this sample, you can use the api key `special-key` to test the authorization filters.",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "apiteam@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "127.0.0.1:8000",
    "basePath": "/api",
    "schemes": [
        "http"
    ],
    "paths": {
...
        "/v1/start-shift": {
            "get": {
                "tags": [
                    "Shifts"
                ],
                "description": "Start driver shift: make him in duty and capture start date",
                "operationId": "startShift",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation returns success message",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Server Error"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
        "/v1/end-shift": {
            "get": {
                "tags": [
                    "Shifts"
                ],
                "description": "End driver shift: make him off duty and capture end date",
                "operationId": "endShift",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Token",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation returns success message",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "SERVER ERROR",
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "default": "Server Error"
                                },
                                "details": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "default": []
                    }
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "default": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}